Fix MSI filename: Use EXE_NAME instead of hardcoded 'Clad'
The OUTPUT_BASE_FILENAME was hardcoded to 'Clad' but should use EXE_NAME variable
which is set to 'Chad' (the actual executable name).

This fixes the error where prepare_assets.sh tries to move:
  build/windows/msi/releasedir/Chad-x64-0.0.29.msi
But the file was actually created as:
  build/windows/msi/releasedir/Clad-x64-0.0.29.msi

Now both will correctly use 'Chad' in the filename.